refactor: UserChatRoom 및 ChatMessage 디렉토리 위치 변경
