+++ src/main/java/space/space_spring/domain/chat/chatting/model/response/ChatMessageResponse.java
@@ -3,7 +3,7 @@ package space.space_spring.domain.chat.chatting.model.response;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import space.space_spring.entity.document.ChatMessage;
+import space.space_spring.domain.chat.chatting.model.document.ChatMessage;
 import space.space_spring.entity.enumStatus.ChatMessageType;
 
 import java.util.HashMap;
